moved MovieCreator component to a extern file in components folder.
